requirements.txt
openai
tenacity
tqdm
numpy
typing
tiktoken
concurrent
ipykernel
pandas
dotenv
psycopg2-binary